menambahkan project 10 dan memperbaiki beberapa nomor

+++ project05.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-void BanyakPecahan(long uang, long pecahan[]){
+void BanyakPecahan(int uang, int pecahan[]){
     int lembar;
 
     for(int i = 0; i < 9; i++){
@@ -14,8 +14,8 @@ void BanyakPecahan(long uang, long pecahan[]){
 }
 
 int main(){
-    long uang;
-    long pecahan[9] = {50000, 20000, 10000, 5000, 2000, 1000, 500, 200, 100};
+    int uang;
+    int pecahan[9] = {50000, 20000, 10000, 5000, 2000, 1000, 500, 200, 100};
 
     cout << "Masukkan jumlah uang kembali: ";
     cin >> uang;
